pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbsdk"
dynamic = ["version"]
requires-python = ">=3.9"
dependencies = [
    "faker>=15.0.0",
    "exrex>=0.11.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "idna>=3.4",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]
llm = ["openai>=1.0.0"]
release = ["build>=1.0.0", "twine>=4.0.0"]

[tool.setuptools.dynamic]
version = {attr = "data_builder.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["data_builder*"]
exclude = ["data_builder.tests*"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*predicate.*安全风险.*:UserWarning",
    "ignore:.*无法解析 constraint predicate.*:UserWarning",
]
